Add the Autonomous GUI
This commit finally adds the GUI #SixMonths

I'm not to sure how to run it in Eclipse, but i was able to in IntelliJ.

It works by just dragging the mouse over the gamefield to create paths,
then the user can save them so that the robot can load them later.

I still have to make the minimum line length configurable, but I really
just wanted to put it on GitHub already so I'll do that some other time.

Enjoy :)