Remove all wackness in the repo
Yo. All I can say is: commiting commented-out code to the main
repository should be an instacut from the team, maybe even from life.

I just wanted a final refactor commit before I start making major
changes to the GUI. I'm going to change the way input works so that the
user just clicks on points instead of dragging lines to create the
paths. This will probably involve refactoring what the Path class is so
that it's not strictly for javafx.scene.shape.Line objects.

It'll be valid.