Add new input mode to Auto GUI
There is now a new input mode in the GUI in which instead of dragging to
generate the paths, the user can click on the points that make up the
line instead. This allows the user to make the lines as detailed as they
want it to be. The user can decide (using a new ChoiceBox) if they want
to enable the new mode or the old one (named "Click mode" and "Drag
mode", respectively).

The Path class has been remade so that it is not only made for
javafx.scene.shape.Line objects. It only stores the coordinate points
now. This means that LinePlus is no longer needed so it was removed.

The status bar has new stats that it tracks to be more useful.