Stop minimumLineLength from allowing letters
Through some error on my part, minimumLineLength allowed letters in the
TextField again, this has been fixed.

The Path class is now based on Integers instead of Doubles because it is
impossible to have a double coordinate value. The class also has an
implemented toString method for testing + a slight JavaDocs update.