Add support for realtime mapping of robot
I added a new package in auto_gui.realtime that handles the new
"Map Mode". Map Mode allows an object to be tracked when given
coordinate values through a java.net.Socket connection to the Server. I
also made a Client class that can be used by the object to track.

In other news, I fixed a bug related to highlighting lines when you try
to delete/save them, and I updated the map file, but this is not the
final one. The next time I update it I will use the final one that is
scaled like the actual game field.